requirements.txt
gunicorn
slack-sdk==3.23.0
flask==3.0.0
python-dotenv==1.0.0
pymongo==4.6.0
requests==2.22.0
requests-unixsocket==0.2.0
mysql-connector-python==8.1.0
PyYAML==5.3.1